Updating discord version and adding new commands
