+++ commands/ban.js
