+++ commands/movemessage.js
