+++ commands/mute.js
