+++ commands/say.js
@@ -1,29 +1,31 @@
 module.exports = {
-  run: (client, message, args) => {
-    if(message.member.permissions.missing(['MANAGE_MESSAGES'])) return message.channel.send('> Você não pode usar esse comando!');
+	run: (client, message, args) => {
+		if (!message.member.hasPermission(['MANAGE_MESSAGES', 'ADMINISTRATOR'])) { return message.channel.send('> Você não pode usar esse comando!'); }
 
-    let msg;
-        let textChannel = message.mentions.channels.first()
-        message.delete()
+		let argsresult;
+		const mChannel = message.mentions.channels.first();
 
-        if(textChannel) {
-            msg = args.slice(1).join(" ");
-            textChannel.send(msg)
-        } else {
-            msg = args.join(" ");
-            message.channel.send(msg)
-        }
-      },
+		message.delete();
+		if (mChannel) {
+			argsresult = args.slice(1).join(' ');
+			mChannel.send(argsresult);
+		}
+		else {
+			argsresult = args.join(' ');
+			message.channel.send(argsresult);
+		}
+	},
 
-  conf: {},
+	conf: {},
 
-  get help () {
-    return {
-      name: 'say',
-      category: 'Moderação',
-      description: 'Faz o bot enviar tal mensagem.',
-      usage: 'say',
-    }
-  }
+	get help() {
+		return {
+			name: 'say',
+			category: 'Moderação',
+			description: 'Faz o bot enviar tal mensagem.',
+			usage: 'say',
+			admin: true,
+		};
+	},
 
-}
+};
