+++ commands/slowmode.js
